+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [251](https://github.com/vegaprotocol/data-node/pull/251) - Updated proto and core and added support for the new events (state var and network limits)
 - [289](https://github.com/vegaprotocol/data-node/pull/289) - Add support for pagination of delegations
 - [254](https://github.com/vegaprotocol/data-node/pull/254) - Move to `ghcr.io` container registry
+- [290](https://github.com/vegaprotocol/data-node/pull/290) - Update pegged orders offset
 
 
 ### 🐛 Fixes